refactor: convert rooms endpoints to TS
