+++ apps/meteor/app/api/server/index.ts
@@ -30,8 +30,7 @@ import './v1/permissions';
 import './v1/presence';
 import './v1/push';
 import './v1/roles';
-import './v1/rooms.js';
-import './v1/rooms.ts';
+import './v1/rooms';
 import './v1/settings';
 import './v1/stats';
 import './v1/subscriptions';
